* Moved logic for  populate tweet list in fragment
